Use long for URN (as it may be a negative timestamp)

+++ app/src/main/java/com/soundcloud/android/playlists/PostedPlaylistMapper.java
@@ -27,7 +27,7 @@ public class PostedPlaylistMapper extends RxResultMapper<PropertySet> {
     }
 
     private Urn readSoundUrn(CursorReader cursorReader) {
-        return Urn.forPlaylist(cursorReader.getInt(BaseColumns._ID));
+        return Urn.forPlaylist(cursorReader.getLong(BaseColumns._ID));
     }
 
 }
